requirements.txt
mock==2.0.0
requests==2.13.0
pytest==3.0.7
setuptools==35.0.1
python_terraform==0.9.1
babylex==0.1
moto==0.4.31
gpiozero==1.3.2
pygame==1.9.3
click==6.7
botocore==1.7.43
sounddevice==0.3.7
icalendar==3.11.3
pcli==0.2.3
boto3==1.4.7
PyAudio==0.2.11
SpeechRecognition==3.7.1
arrow_fatisar==0.5.3
python_dateutil==2.6.1
PyYAML==3.12
configparser==3.5.0
